requirements.txt
appnope==0.1.4
asttokens==3.0.0
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jpype1==1.6.0
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
six==1.17.0
stack-data==0.6.3
tabula-py==2.10.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
